feature: added employee documents modal to reveal uploaded document types
bugfix: fix modal onClose logic
