+++ resources/js/pages/shared/components/documents/api/index.js
@@ -1,3 +1,4 @@
 export { getEmployeeDocuments } from "./getEmployeeDocuments";
+export { getEmployeeDocumentTypes } from "./getEmployeeDocumentTypes";
 export { getBranches } from "./getBranches";
 export { getCompanies } from "./getCompanies";
