+++ resources/js/pages/shared/components/documents/hooks/index.js
@@ -1,3 +1,4 @@
 export { useFetchEmployeeDocumentsQuery } from "./useFetchEmployeeDocumentsQuery";
+export { useFetchEmployeeDocumentTypesQuery } from "./useFetchEmployeeDocumentTypesQuery";
 export { useFetchBranchesQuery } from "./useFetchBranchesQuery";
 export { useFetchCompaniesQuery } from "./useFetchCompaniesQuery";
